+++ test/test_cluda/test_struct.py
@@ -5,6 +5,39 @@ import reikna.cluda as cluda
 import reikna.cluda.dtypes as dtypes
 
 
+def pytest_generate_tests(metafunc):
+
+    if 'dtype_to_adjust' in metafunc.funcargnames:
+
+        vals = []
+        ids = []
+
+        # numpy itemsize == 9, but on device it will be aligned to 4, so the total size will be 12
+        dtype = numpy.dtype([('val1', numpy.int32), ('val2', numpy.int32), ('pad', numpy.int8)])
+        vals.append(dtype)
+        ids.append("small_pad")
+
+        dtype_nested = numpy.dtype([
+            ('val1', numpy.int32), ('pad', numpy.int8)])
+        dtype = numpy.dtype([
+            ('val1', numpy.int32),
+            ('val2', numpy.int16),
+            ('nested', dtype_nested)])
+        vals.append(dtype)
+        ids.append("nested")
+
+        dtype_nested = numpy.dtype(dict(
+            names=['val1', 'pad'],
+            formats=[numpy.int8, numpy.int8]))
+        dtype = numpy.dtype(dict(
+            names=['pad', 'struct_arr', 'regular_arr'],
+            formats=[numpy.int32, numpy.dtype((dtype_nested, 2)), numpy.dtype((numpy.int16, 3))]))
+        vals.append(dtype)
+        ids.append("nested_array")
+
+        metafunc.parametrize('dtype_to_adjust', vals, ids=ids)
+
+
 def extract_field(arr, path):
     if len(path) == 0:
         return arr
@@ -12,10 +45,62 @@ def extract_field(arr, path):
         return extract_field(arr[path[0]], path[1:])
 
 
+def get_offsets_from_device(thr, dtype):
+    """
+    Returns a new data type object with the same fields as in ``dtype``
+    and the field offsets set by the compiler in ``thr``.
+    All existing offsets in ``dtype`` are ignored.
+    """
+
+    if dtype.names is None:
+        return dtype
+
+    struct = dtypes.ctype_module(dtype, ignore_alignment=True)
+
+    program = thr.compile(
+    """
+    #define my_offsetof(type, field) ((size_t)(&((type *)0)->field))
+
+    KERNEL void test(GLOBAL_MEM int *dest)
+    {
+      %for i, name in enumerate(dtype.names):
+      dest[${i}] = my_offsetof(${struct}, ${name});
+      %endfor
+      dest[${len(dtype.names)}] = sizeof(${struct});
+    }
+    """, render_kwds=dict(
+        struct=struct,
+        dtype=dtype))
+
+    offsets = thr.array(len(dtype.names) + 1, numpy.int32)
+    test = program.test
+    test(offsets, global_size=1)
+    offsets = offsets.get()
+
+    # Casting to Python ints, becase numpy ints as dtype offsets make it unhashable.
+    offsets = [int(offset) for offset in offsets]
+
+    adjusted_dtypes = [
+        get_offsets_from_device(thr, dtype.fields[name][0])
+        for name in dtype.names]
+
+    return numpy.dtype(dict(
+        names=dtype.names,
+        formats=adjusted_dtypes,
+        offsets=offsets[:-1],
+        itemsize=offsets[-1]))
+
+
+def test_adjust_offsets(thr, dtype_to_adjust):
+    adjusted_dtype = dtypes.adjust_offsets(dtype_to_adjust)
+    empyric_dtype = get_offsets_from_device(thr, dtype_to_adjust)
+    assert adjusted_dtype == empyric_dtype
+
+
 def check_struct_fill(thr, dtype):
     """
     Fill every field of the given ``dtype`` with its number and check the results.
-    This helps to detect issues with alignment in the struct.
+    This helps to detect issues with offsets in the struct.
     """
     struct = dtypes.ctype_module(dtype)
 
@@ -47,7 +132,7 @@ def check_struct_fill(thr, dtype):
         assert (extract_field(a, path) == i).all()
 
 
-def test_hardcoded_alignment(thr):
+def test_hardcoded_offsets(thr):
     """
     Test the correctness of alignment for an explicit set of field offsets.
     """
@@ -80,7 +165,7 @@ def test_adjusted_alignment(thr):
         ('val2', numpy.int16),
         ('nested', dtype_nested)])
 
-    dtype = dtypes.adjust_alignment(thr, dtype)
+    dtype = dtypes.adjust_offsets(dtype)
 
     check_struct_fill(thr, dtype)
 
@@ -97,7 +182,7 @@ def test_nested_array(thr):
         names=['pad', 'struct_arr', 'regular_arr'],
         formats=[numpy.int32, numpy.dtype((dtype_nested, 2)), numpy.dtype((numpy.int16, 3))]))
 
-    dtype = dtypes.adjust_alignment(thr, dtype)
+    dtype = dtypes.adjust_offsets(dtype)
     struct = dtypes.ctype_module(dtype)
 
     program = thr.compile(